scripts/requirements.txt
numpy
matplotlib
pandas
sklearn
opencv-python
flask
lapjv
pillow